+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.11.1
 - Change theme painter default to module.cavern theme.
+- Start playing any video textures for background or floor.
 
 # 0.11.0
 - Add theme painter / eraser tool, that allows adding "theme areas" on top of the main map theme.